requirements.txt
apted==1.0.3
beautifulsoup4==4.12.3
Distance==0.1.3
lxml==5.2.1
matplotlib==3.7.3
numpy==1.24.3
opencv-python-headless==4.10.0.84
pandas==2.0.3
Pillow==9.5.0
Polygon3==3.0.9.1
protobuf==3.20.0
prov==2.0.0
scikit-image==0.20.0
scipy==1.9.1
shapely==2.0.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.65.0
transformers==4.28.1

